backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0  # 如果使用SQLite
# psycopg2-binary>=2.9.9  # 如果使用PostgreSQL 